+++ src/textual/events.py
@@ -563,7 +563,8 @@ class Blur(Event, bubble=False):
 class AppFocus(Event, bubble=False):
     """Sent when the app has focus.
 
-    Used by textual-web.
+    Only available when running within a terminal that supports `FocusIn`,
+    or when running via textual-web.
 
     - [ ] Bubbles
     - [ ] Verbose
@@ -573,7 +574,8 @@ class AppFocus(Event, bubble=False):
 class AppBlur(Event, bubble=False):
     """Sent when the app loses focus.
 
-    Used by textual-web.
+    Only available when running within a terminal that supports `FocusOut`,
+    or when running via textual-web.
 
     - [ ] Bubbles
     - [ ] Verbose